+++ src/zcl/zclHeader.ts
@@ -1,0 +1,10 @@
+import {FrameControl} from './definition';
+
+interface ZclHeader {
+    frameControl: FrameControl;
+    manufacturerCode: number;
+    transactionSequenceNumber: number;
+    commandIdentifier: number;
+}
+
+export default ZclHeader;